requirements.txt
aiohttp==3.8.1
requests
pyslack
raven
Django==4.0
dj-database-url
whitenoise
sqlparse
python-dateutil
dj-static
mimeparse
django-avatar
django-bootstrap3
django-debug-toolbar
django-extensions
django-extra-views
django-stronghold
gunicorn
honcho
psycopg2
pytz
redis
django-redis
django-tastypie==0.14.4
django-braces